working on frontend user listing
