+++ react-app/src/components/Listing/UserListing.css
@@ -1,0 +1,62 @@
+.owner-items-outer-container{
+    border: 1px solid hotpink;
+    box-sizing: border-box;
+    color: rgb(34,34,34);
+    font-size: 14px;
+    height: 1138px;
+    line-height: 19.6px;
+    margin: 0px 0px;
+    padding-bottom: 12px;
+    padding-left: 18px;
+    padding-right: 18px;
+    padding-top: 30px;
+    text-size-adjust: 100%;
+    text-align: left;
+    width: 1031px;
+    display: block;
+}
+.onwer-listing{
+    /* border: 1px solid blueviolet; */
+    box-sizing: border-box;
+    color: rgb(34,34,34);
+    font-size: 32px;
+    font-weight: 300;
+    height: 36px;
+    letter-spacing: 0.5px;
+    line-height: 36px;
+    text-align: center;
+    text-size-adjust: 100%;
+    width: 995px;
+
+}
+.owner-listing-subtitle{
+    /* border: 1px solid black; */
+     box-sizing: border-box;
+    color: rgb(34,34,34);
+    font-size: 16px;
+    font-weight: 400;
+    height: 84px;
+    line-height: 28px;
+    margin-bottom: 30px;
+    margin-left: 250px;
+    margin-right: 250px;
+    margin-top:0px ;
+    padding-left: 18px;
+    padding-right: 18px;
+    text-align: center;
+    width: 500px;
+}
+.owner-items-inner-container{
+    border: 1px solid rgb(128, 128, 128,0.5);
+    box-sizing: border-box;
+    color: rgb(34,34,34);
+    display: block;
+    font-size: 14px;
+    height: 962px;
+    line-height: 19.6px;
+    margin: 0px 0px;
+    padding: 36px;
+    text-align: left;
+    text-size-adjust: 100%;
+    /* width: 1031px; */
+}
